Move Launcher into core package
This is because it is not of general interest to benchmark writers.
